Soft Lock Fix
I think the issue with the turns was indeed related to interaction between preventing advancing (cannot move forward if there is an injury) and the current tab it was in. I made a little bit of changes to the tabs and I was not able to replicate the soft lock error, we might be able to close that issue. If it reoccurs then we can reopen it, but as for now it seems as if this works.